+++ plugin-gestures/src/main/java/com/mapbox/maps/plugin/gestures/GesturesPluginImpl.kt
@@ -99,10 +99,18 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
   private val scheduledAnimators = ArrayList<ValueAnimator>()
   private var gesturesInterpolator = LinearOutSlowInInterpolator()
 
-  // needed most likely for devices with API <= 23 only
-  // duration = 0 will still make animation end / cancel not immediately
-  // this may cause cancelling some easeTo animation without single camera update
+  // Devices with API <= 23 animating with duration = 0 will send initial value from Looper with
+  // some small delay as a result.
+  // Hence multiple immediate animations sent close to each other may cancel previous ones.
+  // This becomes visible with move gesture since animation steps are relative to each other.
+  // So we defer move distances that are not applied to apply later (when immediate ease will end).
   private var immediateEaseInProcess = false
+  private var deferredMoveDistanceX: Float = 0f
+  private var deferredMoveDistanceY: Float = 0f
+  private var deferredZoomBy: Double = 0.0
+  private var deferredRotate: Double = 0.0
+  private var deferredShove: Double = 0.0
+
   private fun easeToImmediately(
     camera: CameraOptions,
     actionAfter: (() -> Unit)? = null
@@ -598,6 +606,7 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
   internal fun handleScale(detector: StandardScaleGestureDetector): Boolean {
     // in order not to mess up initial anchor values
     if (!internalSettings.simultaneousRotateAndPinchToZoomEnabled && immediateEaseInProcess) {
+      deferredZoomBy += calculateZoomBy(detector)
       return true
     }
     val focalPoint = getScaleFocalPoint(detector)
@@ -628,8 +637,7 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
         }
       )
     } else {
-      val zoomBy =
-        ln(detector.scaleFactor.toDouble()) / ln(PI / 2) * ZOOM_RATE.toDouble() * internalSettings.zoomAnimationAmount.toDouble()
+      val zoomBy = calculateZoomBy(detector)
       if (internalSettings.simultaneousRotateAndPinchToZoomEnabled) {
         val zoom = cameraAnimationsPlugin.createZoomAnimator(
           cameraAnimatorOptions(mapCameraManagerDelegate.cameraState.zoom + zoomBy) {
@@ -655,13 +663,14 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
       } else {
         easeToImmediately(
           CameraOptions.Builder()
-            .zoom(mapCameraManagerDelegate.cameraState.zoom + zoomBy)
+            .zoom(mapCameraManagerDelegate.cameraState.zoom + zoomBy + deferredZoomBy)
             .anchor(focalPoint)
             .build(),
           actionAfter = {
             onScaleAnimationEnd(detector)
           }
         )
+        deferredZoomBy = 0.0
       }
     }
     return true
@@ -675,6 +684,10 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
 
   internal fun handleScaleBegin(detector: StandardScaleGestureDetector): Boolean {
     quickZoom = detector.pointersCount == 1
+    // Reset deferred animation values here in onBegin, ignores the last possible defererred value
+    // before the onEnd. However it's not noticeable since onEnd animates using velocity.
+    // Allows to keep things simple.
+    deferredZoomBy = 0.0
 
     if (quickZoom) {
       if (!internalSettings.quickZoomEnabled) {
@@ -875,12 +888,14 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
   ): Boolean {
     // in order not to mess up initial anchor values
     if (!internalSettings.simultaneousRotateAndPinchToZoomEnabled && immediateEaseInProcess) {
+      deferredRotate += rotationDegreesSinceLast
       return true
     }
     // Calculate map bearing value
     val currentBearing = mapCameraManagerDelegate.cameraState.bearing
     rotateCachedAnchor = cameraAnimationsPlugin.anchor
-    val bearing = currentBearing + rotationDegreesSinceLast
+    val bearing = currentBearing + rotationDegreesSinceLast + deferredRotate
+    deferredRotate = 0.0
     val focalPoint = getRotateFocalPoint(detector)
     // Rotate the map
     if (internalSettings.simultaneousRotateAndPinchToZoomEnabled) {
@@ -928,6 +943,11 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
     if (!internalSettings.rotateEnabled) {
       return false
     }
+
+    // Reset deferred animation values here in onBegin, ignores the last possible defererred value
+    // before the onEnd. However it's not noticeable since onEnd animates using velocity.
+    // Allows to keep things simple.
+    deferredRotate = 0.0
 
     val deltaSinceLast = abs(detector.deltaSinceLast)
     val currTime = detector.currentEvent.eventTime.toDouble()
@@ -998,6 +1018,11 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
       return false
     }
 
+    // Reset deferred animation values here in onBegin, ignores the last possible defererred value
+    // before the onEnd. However it's not noticeable since onEnd animates using velocity.
+    // Allows to keep things simple.
+    deferredShove = 0.0
+
     cancelTransitionsIfRequired()
 
     // disabling move gesture during shove
@@ -1012,9 +1037,14 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
     detector: ShoveGestureDetector,
     deltaPixelsSinceLast: Float
   ): Boolean {
+    if (immediateEaseInProcess) {
+      deferredShove += deltaPixelsSinceLast
+      return true
+    }
     // Get pitch value (scale and clamp)
     var pitch = mapCameraManagerDelegate.cameraState.pitch
-    val optimizedPitch = pitch - (SHOVE_PIXEL_CHANGE_FACTOR * deltaPixelsSinceLast).toDouble()
+    val optimizedPitch = pitch - (SHOVE_PIXEL_CHANGE_FACTOR * (deltaPixelsSinceLast + deferredShove))
+    deferredShove = 0.0
     pitch = clamp(optimizedPitch, MINIMUM_PITCH, MAXIMUM_PITCH)
     easeToImmediately(
       CameraOptions.Builder().pitch(pitch).build(),
@@ -1331,6 +1361,12 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
       return false
     }
 
+    // Reset deferred animation values here in onBegin, ignores the last possible defererred value
+    // before the onEnd. However it's not noticeable since onEnd animates using velocity.
+    // Allows to keep things simple.
+    deferredMoveDistanceX = 0f
+    deferredMoveDistanceY = 0f
+
     cancelTransitionsIfRequired()
     notifyOnMoveBeginListeners(detector)
     return true
@@ -1362,6 +1398,13 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
     val roundtripPoint = mapCameraManagerDelegate.pixelForCoordinate(coordinate)
     return (roundtripPoint.y >= point.y + reprojectErrorMargin)
   }
+
+  @VisibleForTesting(otherwise = PRIVATE)
+  internal fun calculateZoomBy(standardScaleGestureDetector: StandardScaleGestureDetector) =
+    ln(standardScaleGestureDetector.scaleFactor.toDouble()) /
+      ln(PI / 2) *
+      ZOOM_RATE.toDouble() *
+      internalSettings.zoomAnimationAmount.toDouble()
 
   internal fun handleMove(
     detector: MoveGestureDetector,
@@ -1401,10 +1444,17 @@ class GesturesPluginImpl : GesturesPlugin, GesturesSettingsBase, MapStyleObserve
         dragInProgress = true
         mapCameraManagerDelegate.dragStart(ScreenCoordinate(fromX, fromY))
       }
+      if (immediateEaseInProcess) {
+        deferredMoveDistanceX += distanceX
+        deferredMoveDistanceY += distanceY
+        return true
+      }
       val resolvedDistanceX =
-        if (internalSettings.isScrollHorizontallyLimited()) 0.0 else distanceX.toDouble()
+        if (internalSettings.isScrollHorizontallyLimited()) 0.0 else (distanceX.toDouble() + deferredMoveDistanceX)
       val resolvedDistanceY =
-        if (internalSettings.isScrollVerticallyLimited()) 0.0 else distanceY.toDouble()
+        if (internalSettings.isScrollVerticallyLimited()) 0.0 else (distanceY.toDouble() + deferredMoveDistanceY)
+      deferredMoveDistanceX = 0f
+      deferredMoveDistanceY = 0f
 
       val toX = fromX - resolvedDistanceX
       val toY = fromY - resolvedDistanceY